Change canvas.toBuffer() params, for big speedup
See "docs" at https://github.com/Automattic/node-canvas/issues/306 and
https://www.w3.org/TR/PNG/

The gist is: node-canvas is horrid because it doesn't give access to the
raw image. Instead, it outputs PNG or JPEG or SVG or PDF. JPEG would be
_okay_ ... but PNG would be nicer. Luckily, node-canvas offers a couple
of options.

First, PNG_FILTER_NONE: this prevents libpng from modifying each
scanline, as it otherwise would. (Coincidentally, it makes our PNGs
compress better, probably because zlib is adept at compressing solid
blocks of color.) It saves a lot of time.

Second, zlib compression of 1: this is much, much faster than the
default of 6.